requirements.txt
fastapi[standard]>=0.112.0,<0.113
SQLAlchemy>=2.0.32,<2.1
passlib>=1.7.4,<1.8
bcrypt>=4.2.0,<4.3
python-jose>=3.3.0,<3.4
python-dotenv>=1.0.1,<1.1
pydantic>=2.8.2,<2.9
gunicorn>=24.1.1,<24.2
uvicorn>=0.40.0,<0.41